rearrange the public API
I want bin/inspect to output info about the file, without printing the
inflated output.

I was particularly interested in a quick way to see the number of blocks
in a file, and what encoding method they use. That required a way to
iterate over the blocks from outside Deflate::File.